publisher/requirements.txt
EbookLib==0.18
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
six==1.16.0
